requirements.txt
boto3==1.34.127
botocore==1.34.127
fastapi==0.111.1
passlib==1.7.4
pydantic==2.8.2
PyJWT==2.8.0
PyJWT==2.8.0
uvicorn==0.30.1
